+++ src/api/pki/responses.rs
@@ -1,3 +1,5 @@
+use std::collections::HashMap;
+
 use serde::{Deserialize, Serialize};
 
 /// Response from executing
@@ -94,13 +96,18 @@ pub struct ReadRoleResponse {
     pub allow_localhost: bool,
     pub allow_subdomains: bool,
     pub allow_token_displayname: bool,
+    pub allow_wildcard_certificates: bool,
     pub allowed_domains: Vec<String>,
     pub allowed_domains_template: bool,
     pub allowed_other_sans: Vec<String>,
     pub allowed_serial_numbers: Vec<String>,
     pub allowed_uri_sans: Vec<String>,
+    pub allowed_uri_sans_template: bool,
+    // todoi: added in 1.14.2 but works with 1.13.3
+    pub allowed_user_ids: Vec<String>,
     pub basic_constraints_valid_for_non_ca: bool,
     pub client_flag: bool,
+    pub cn_validations: Vec<String>,
     pub code_signing_flag: bool,
     pub country: Vec<String>,
     pub email_protection_flag: bool,
@@ -108,12 +115,14 @@ pub struct ReadRoleResponse {
     pub ext_key_usage: Vec<String>,
     pub ext_key_usage_oids: Vec<String>,
     pub generate_lease: bool,
+    pub issuer_ref: String,
     pub key_bits: u64,
     pub key_type: String,
     pub key_usage: Vec<String>,
     pub locality: Vec<String>,
     pub max_ttl: u64,
     pub no_store: bool,
+    pub not_after: String,
     pub not_before_duration: u64,
     pub organization: Vec<String>,
     pub ou: Vec<String>,
@@ -122,10 +131,12 @@ pub struct ReadRoleResponse {
     pub province: Vec<String>,
     pub require_cn: bool,
     pub server_flag: bool,
+    pub signature_bits: u16,
     pub street_address: Vec<String>,
     pub ttl: u64,
     pub use_csr_common_name: bool,
     pub use_csr_sans: bool,
+    pub use_pss: bool,
 }
 
 /// Response from executing
@@ -154,4 +165,26 @@ pub struct SignIntermediateResponse {
 pub struct SignSelfIssuedResponse {
     pub certificate: String,
     pub issuing_ca: String,
+}
+
+/// Response from executing
+/// [ReadIssuerCertificateRequest][crate::api::pki::requests::ReadIssuerCertificateRequest]
+#[derive(Deserialize, Debug, Serialize)]
+pub struct ReadIssuerCertificateResponse {
+    pub certificate: String,
+    pub ca_chain: Vec<String>,
+    pub issuer_id: String,
+    pub issuer_name: String,
+}
+
+/// Response from executing
+/// [ImportIssuerRequest][crate::api::pki::requests::ImportIssuerRequest]
+#[derive(Deserialize, Debug, Serialize)]
+pub struct ImportIssuerResponse {
+    pub imported_issuers: Option<Vec<String>>,
+    pub imported_keys: Option<Vec<String>>,
+    // todoi: added in 1.14.0
+    pub existing_issuers: Option<Vec<String>>,
+    pub existing_keys: Option<Vec<String>>,
+    pub mapping: Option<HashMap<String, String>>,
 }