+++ CHANGELOG.md
@@ -8,6 +8,12 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Added
+
+- Read issuer's certificate (see `issuer::read`)
+- Import issuer (see `issuer::import`)
+- Delete issuer (see `issuer::delete`)
+
 ## [0.7.0] - 2023-03-25
 
 ### Added
@@ -109,7 +115,7 @@ and this project adheres to
 - Support for getting server status
 - Support for using CA certificates in requests
 - Generic login method for clients to easily obtain tokens
-- Methods for isting available auth methods on a server
+- Methods for listing available auth methods on a server
 - Additional support for automating OIDC logins
 
 ## [0.3.0] - 2021-08-30